Fix ps binary path and arguments for Solaris 11
On Solaris 10 the path to BSD ps(1B) is /usr/ucb/ps. On Solaris 11 the
BSD compatibility tools are not installed by default, and ps(1) has been
updated to detect BSD-style arguments and emulate the corresponding
behaviour automatically. Therefore path can be changed to /bin/ps for
Solaris 11 versus /usr/ucb/ps for Solaris 10.

Also changed the way Facter fact, kernel, is obtained as using a Ruby
symbol for the fact resulted in the literal string "kernel" as opposed
to the anticipated fact value, i.e. "SunOS".