Add support for multiple base installations of same Fmw
The current version of fmw, doesn't support the installation of the same middleware more than once. If you try, you get duplicate resources.  If however you want to have a middleware home for SOA and one for BAM and apply different patches to these, you'll need to do just that. This PR add's support for installing multiple homes with the same version. You will have to use different titles to the fmw to do this.  After this you can use opatch to differentiate the middleware homes