Improve wls_server 'arguments' property
The original code just took arguments as a simple string.  The code was then improved to
accept arrays too.  Arrays were flattened to newline separated strings before being passed
to the template.

A few issues with this.
Weblogic actually preferred server arguments as space separated strings.
In weblogic 12.1, if you use the web interface to enter arguments one per line, they get
converted to space separated when you hit 'save'.

If an array was used to set arguments, output from puppet resource was as the newline separated
string.  This looked a bit nasty.  It would have been much nicer to have puppet return an array
of arguments no matter how they were created.

The order of the arguments would be taken into account when comparing is and should
values.  Things that should probably have been idempotent weren't.

This commit retains the ability to specify the arguments as either a string or an
array.  But now, order is not important.
['arg1','arg2'] is considered the same as 'arg2 arg1'
If puppet does need to update the property, the puppet output is customised to say
which arguments (if any) are being removed, and which are being added.